Reduce bundle size by importing icons individually and effects asynchronously.
