+++ models/vote.js
@@ -1,13 +1,7 @@
-const { ASSETS_URL } = process.env
-const stateNames = require('datasets-us-states-abbr-names')
 const {
-  api, avatarURL, combineEffects, combineEffectsInSeries, escapeHtml,
-  makePoint, possessive, preventDefault, waitEffects
+  combineEffects, combineEffectsInSeries, preventDefault
 } = require('../helpers')
 const { fetchMeasure } = require('../effects/measure')
-const { updateNameAndAddress } = require('../effects/user')
-const { signIn } = require('../effects/session')
-const { changePageTitle } = require('../effects/page')
 const { logEndorsed } = require('../effects/analytics')
 
 module.exports = (event, state) => {
@@ -17,27 +11,19 @@ module.exports = (event, state) => {
         case '/legislation/:shortId/votes/:voteId':
         case '/nominations/:shortId/votes/:voteId':
         case '/:username/:shortId/votes/:voteId':
-          const vote = state.votes && state.votes[state.location.params.voteId]
-          const commentsLoaded = vote && vote.replies
-          const backersLoaded = vote && vote.backers
           return [{
             ...state,
-            loading: { ...state.loading, page: !state.votes[event.voteId], backers: !backersLoaded, comments: !commentsLoaded },
-            backersFilterQuery: decodeURI(state.location.query.filter || ''),
+            loading: { ...state.loading, page: !state.votes[event.voteId] },
             votes: {
               ...state.votes,
               [event.voteId]: {
                 ...state.votes[event.voteId],
-                showQuestionForm: state.location.hash === 'endorsement-question',
                 expanded: state.location.query.show_more === 'true',
               },
             },
-          }, waitEffects([
-            combineEffectsInSeries([
-              fetchMeasure(state.location.params.shortId, state.offices, state.user),
-              fetchVote(state.location.params.voteId, state.user),
-            ]),
-            fetchVoteReplies(state.location.params.voteId, state.user),
+          }, combineEffectsInSeries([
+            importEffect('fetchVote', state.location.params.voteId, state.user),
+            fetchMeasure(state.location.params.shortId, state),
           ])]
         default:
           return [state]
@@ -67,23 +53,14 @@ module.exports = (event, state) => {
             reported: true,
           },
         },
-      }, combineEffects([preventDefault(event.event), reportVote(event.vote, state.user)])]
+      }, combineEffects([preventDefault(event.event), importEffect('reportVote', event.vote, state.user)])]
     case 'vote:endorsed':
       // TODO cannot change state here otherwise iframes in questions will re-render and Chrome will dismiss the confirm dialog whenever an iframe loads.
       return [state, combineEffectsInSeries([
         preventDefault(event.event),
-        endorse(event.vote, state.user, event.measure, event.is_public),
-        typeof event.name !== 'undefined' && state.user && updateNameAndAddressFromEndorsement(event, state.user),
+        importEffect('endorse', event.vote, state.user, event.measure, event.is_public),
+        typeof event.name !== 'undefined' && state.user && importEffect('updateNameAndAddressFromEndorsement', event, state.user),
         fetchMeasure(event.vote.measure.short_id, state),
-        logEndorsed,
-      ])]
-    case 'vote:endorsedFromSignupForm':
-      return [{
-        ...state,
-        loading: { ...state.loading, vote: true, endorsedFromSignupForm: true },
-      }, combineEffectsInSeries([
-        preventDefault(event.event),
-        signupAndEndorse(event, state),
         logEndorsed,
       ])]
     case 'vote:unendorsed':
@@ -92,36 +69,23 @@ module.exports = (event, state) => {
         loading: { ...state.loading, vote: true },
       }, combineEffectsInSeries([
         preventDefault(event.event),
-        unendorse(event.vote, state.user),
+        importEffect('unendorse', event.vote, state.user),
         fetchMeasure(event.vote.measure.short_id, state),
       ])]
     case 'vote:changedPrivacy':
       return [state, combineEffectsInSeries([
         preventDefault(event.event),
-        changeVotePrivacy(event.vote, event.public, state.user),
+        importEffect('changeVotePrivacy', event.vote, event.public, state.user),
         fetchMeasure(event.vote.measure.short_id, state),
       ])]
     case 'vote:received':
+    case 'vote:updated':
       if (!event.vote) {
         return [{ ...state, loading: { page: false }, location: { ...state.location, status: 404 } }]
       }
-      const endorsementLocation = endorsementPageTitleAndMeta(state.measures, event.vote, state.location)
       return [{
         ...state,
-        loading: { ...state.loading, page: false, vote: false },
-        location: endorsementLocation,
-        votes: {
-          ...state.votes,
-          [event.vote.id]: {
-            ...state.votes[event.vote.id],
-            ...event.vote,
-          },
-        },
-      }, changePageTitle(endorsementLocation.title)]
-    case 'vote:updated':
-      return [{
-        ...state,
-        loading: { ...state.loading, vote: false, endorsedFromSignupForm: false },
+        loading: { ...state.loading, vote: false },
         votes: {
           ...state.votes,
           [event.vote.id]: {
@@ -130,46 +94,6 @@ module.exports = (event, state) => {
           },
         },
       }]
-    case 'vote:replied':
-      return [{
-        ...state,
-        loading: { ...state.loading, comments: true },
-      }, combineEffects([preventDefault(event.event), reply(event, state.user)])]
-    case 'vote:repliesReceived':
-      return [{
-        ...state,
-        loading: { ...state.loading, comments: false },
-        votes: {
-          ...state.votes,
-          [event.voteId]: {
-            ...state.votes[event.voteId],
-            replies: event.replies,
-          },
-        },
-      }]
-    case 'vote:backersReceived':
-      return [{
-        ...state,
-        loading: { ...state.loading, backers: false },
-        votes: {
-          ...state.votes,
-          [event.voteId]: {
-            ...state.votes[event.voteId],
-            backers: event.backers,
-          },
-        },
-      }]
-    case 'vote:questionFormActivated':
-      return [{
-        ...state,
-        votes: {
-          ...state.votes,
-          [event.vote.id]: {
-            ...state.votes[event.vote.id],
-            showQuestionForm: true,
-          },
-        },
-      }, combineEffects([preventDefault(event.event), scrollQuestionFormIntoView])]
     case 'vote:voted':
       return [{
         ...state,
@@ -179,337 +103,17 @@ module.exports = (event, state) => {
           last_vote_public: event.public,
         } : state.user,
       }, combineEffectsInSeries([
-        vote(event, state.user),
+        preventDefault(event.event),
+        importEffect('vote', event, state.user),
         fetchMeasure(event.measure.short_id, state),
-      ])]
-    case 'vote:backersFilterUpdated':
-      const query = event.event.target ? event.event.target.value : ''
-
-      // Set query in URL bar &filter=${query} after delay
-      debounce(() => {
-        window.history.replaceState({}, '', `${state.location.path}?tab=backers&filter=${query}`)
-      }, 50)
-
-      return [{
-        ...state,
-        backersFilterQuery: query,
-      }]
-    case 'vote:fetchRepliesAndBackers':
-      return [{
-        ...state,
-      }, combineEffects([
-        fetchVoteReplies(state.location.params.voteId, state.user),
-        fetchVoteBackers(state.location.params.voteId),
       ])]
     default:
       return [state]
   }
 }
 
-const vote = ({ event, measure, ...form }, user) => (dispatch) => {
-  if (event) event.preventDefault()
-
-  if (!form.position) {
-    return dispatch({ type: 'error', error: new Error('You must choose a position.') })
-  }
-
-  if (!user) {
-    dispatch({ type: 'cookieSet', key: 'vote_bill_id', value: measure.id })
-    dispatch({ type: 'cookieSet', key: 'vote_bill_short_id', value: measure.short_id })
-    dispatch({ type: 'cookieSet', key: 'vote_position', value: form.position })
-    dispatch({ type: 'cookieSet', key: 'vote_public', value: form.public ? 'true' : '' })
-    dispatch({ type: 'cookieSet', key: 'vote_comment', value: form.comment })
-
-    return dispatch({ type: 'redirected', url: '/join' })
-  }
-
-  return api(dispatch, `/votes?user_id=eq.${user.id}&measure_id=eq.${measure.id}`, {
-    method: measure.vote ? 'PATCH' : 'POST',
-    body: JSON.stringify({
-      user_id: user.id,
-      measure_id: measure.id,
-      position: form.position,
-      root_delegate_id: user.id,
-      delegate_rank: -1,
-      delegate_id: null,
-      delegate_name: null,
-      comment: form.comment || null,
-      public: form.public,
-    }),
-    user,
-  })
-  .then(() => {
-    dispatch({ type: 'measure:voteFormToggled', measure })
-    return api(dispatch, `/votes_detailed?user_id=eq.${user.id}&measure_id=eq.${measure.id}`, { user })
-  })
-  .then(([vote]) => dispatch({ type: 'vote:updated', vote }))
-  .catch((error) => dispatch({ type: 'error', error }))
-}
-
-const reportVote = (vote, user) => (dispatch) => {
-  if (!vote.reported) {
-    return api(dispatch, '/reports', {
-      method: 'POST',
-      body: JSON.stringify({
-        reporter_id: user.id,
-        comment_author_id: vote.user_id,
-        comment_id: vote.id,
-        explanation: null,
-      }),
-      user,
-    })
-    .catch((error) => dispatch({ type: 'error', error }))
-  }
-}
-
-const endorse = (vote, user, measure, is_public = false) => (dispatch) => {
-  const endorsed_vote = !(user && user.id === vote.user_id && vote.comment) && vote.endorsed_vote
-  const { fullname, measure_id, short_id, id: vote_id } = endorsed_vote || vote
-  const position = measure && measure.vote && measure.vote.position
-
-  if (!user) {
-    dispatch({ type: 'cookieSet', key: 'endorsed_vote_id', value: vote_id })
-    dispatch({ type: 'cookieSet', key: 'endorsed_measure_id', value: measure_id })
-    dispatch({ type: 'cookieSet', key: 'endorsed_url', value: `/legislation/${short_id}/votes/${vote_id}` })
-    return dispatch({ type: 'redirected', url: '/join' })
-  }
-
-  if (position) {
-    let confirmation_text = 'You\'ve already '
-    if (measure.vote && measure.vote.position) {
-      confirmation_text += `commented. This will remove your previous comment`
-    } else {
-      confirmation_text += `voted ${position}`
-    }
-    confirmation_text += `. Endorse ${fullname ? possessive(fullname) : 'this'} vote instead?`
-    if (!window.confirm(confirmation_text)) {
-      return
-    }
-  }
-
-  return api(dispatch, `/endorsements?user_id=eq.${user.id}&measure_id=eq.${measure_id}`, {
-    method: position && measure.endorsed ? 'PATCH' : 'POST',
-    body: JSON.stringify({
-      user_id: user.id,
-      vote_id,
-      measure_id,
-      public: is_public,
-    }),
-    user,
-  })
-  .then(() => api(dispatch, `/votes_detailed?id=eq.${vote.id}`, { user }))
-  .then(([vote]) => dispatch({ type: 'vote:updated', vote }))
-  .catch((error) => {
-    if (error.code === 23505) {
-      return api(dispatch, `/endorsements?user_id=eq.${user.id}&measure_id=eq.${measure_id}`, {
-        method: 'PATCH',
-        body: JSON.stringify({
-          user_id: user.id,
-          vote_id,
-          measure_id,
-          public: is_public,
-        }),
-        user,
-      })
-      .then(() => api(dispatch, `/votes_detailed?id=eq.${vote.id}`, { user }))
-      .then(([vote]) => dispatch({ type: 'vote:updated', vote }))
-      .catch((error) => dispatch({ type: 'error', error }))
-    }
-    dispatch({ type: 'error', error })
+const importEffect = (name, ...args) => (dispatch) => {
+  return import('../effects/vote').then((voteEffects) => {
+    return (voteEffects.default || voteEffects)[name].apply(null, args)(dispatch)
   })
 }
-
-const unendorse = (vote, user) => (dispatch) => {
-  if (!user) {
-    return dispatch({ type: 'redirected', url: '/join' })
-  }
-  if (!window.confirm(`Are you sure you want to remove this endorsement?`)) {
-    return
-  }
-  return api(dispatch, `/votes_detailed?user_id=eq.${user.id}&measure_id=eq.${vote.measure_id}`, {
-    method: 'PATCH',
-    body: JSON.stringify({
-      position: 'abstain',
-      delegate_rank: -1,
-      root_delegate_id: user.id,
-      delegate_id: null,
-    }),
-    user,
-  })
-  .then(() => api(dispatch, `/votes_detailed?id=eq.${vote.id}`, { user }))
-  .then(([vote]) => dispatch({ type: 'vote:updated', vote }))
-  .catch((error) => dispatch({ type: 'error', error }))
-}
-
-const changeVotePrivacy = (vote, is_public, user) => (dispatch) => {
-  const endorsed_vote = !(user && user.id === vote.user_id && vote.comment) && vote.endorsed_vote
-  const { measure_id, id: vote_id } = endorsed_vote || vote
-
-  return api(dispatch, `/endorsements?user_id=eq.${user.id}&measure_id=eq.${measure_id}`, {
-    method: 'PATCH',
-    body: JSON.stringify({ user_id: user.id, vote_id, measure_id, public: is_public }),
-    user,
-  })
-  .then(() => api(dispatch, `/votes_detailed?measure_id=eq.${measure_id}&id=eq.${vote_id}`, { user }))
-  .then(([vote]) => dispatch({ type: 'vote:updated', vote }))
-  .catch((error) => dispatch({ type: 'error', error }))
-}
-
-const fetchVote = (id, user) => (dispatch) => {
-  return api(dispatch, `/votes_detailed?id=eq.${id}`, { user })
-    .then(([vote]) => dispatch({ type: 'vote:received', vote }))
-    .catch((error) => dispatch({ type: 'error', error }))
-}
-
-const fetchVoteReplies = (voteId, user) => (dispatch) => {
-  return api(dispatch, `/replies_detailed?vote_id=eq.${voteId}&order=created_at.desc`, { user })
-    .then((replies) => dispatch({ type: 'vote:repliesReceived', voteId, replies }))
-    .catch((error) => dispatch({ type: 'error', error }))
-}
-
-const fetchVoteBackers = (voteId) => (dispatch) => {
-  return api(dispatch, `/endorsements_detailed?vote_id=eq.${voteId}&order=created_at.asc`, { getAllPages: true })
-    .then((backers) => dispatch({ type: 'vote:backersReceived', voteId, backers }))
-    .catch((error) => dispatch({ type: 'error', error }))
-}
-
-const reply = ({ vote, content }, user) => (dispatch) => {
-  const reply = {
-    vote_id: vote.id,
-    user_id: user.id,
-    content,
-  }
-
-  api(dispatch, `/replies`, {
-    method: 'POST',
-    body: JSON.stringify(reply),
-    user,
-  })
-  .then(() => api(dispatch, `/replies_detailed?vote_id=eq.${vote.id}&order=created_at.desc`, { user }))
-  .then((replies) => dispatch({ type: 'vote:repliesReceived', voteId: vote.id, replies }))
-}
-
-const validateNameAndAddressForm = (address, name) => {
-  const name_pieces = name.split(' ')
-
-  if (name_pieces.length < 2) {
-    return Object.assign(new Error('Please enter a first and last name'), { field: 'name' })
-  } else if (name_pieces.length > 5) {
-    return Object.assign(new Error('Please enter only a first and last name'), { field: 'name' })
-  }
-
-  if (!address.match(/ \d{5}/) && (!window.lastSelectedGooglePlacesAddress || !window.lastSelectedGooglePlacesAddress.lon)) {
-    return Object.assign(
-      new Error(`Please use your complete address including city, state, and zip code.`),
-      { field: 'address' }
-    )
-  }
-}
-
-const updateNameAndAddressFromEndorsement = (form, user) => (dispatch) => {
-  const { address, voter_status } = form
-  const error = validateNameAndAddressForm(address, form.name)
-
-  if (error) return dispatch({ type: 'error', error })
-
-  const name_pieces = form.name.split(' ')
-  const first_name = name_pieces[0]
-  const last_name = name_pieces.slice(1).join(' ')
-
-  return updateNameAndAddress({
-    addressData: {
-      address,
-      locality: window.lastSelectedGooglePlacesAddress.locality,
-      administrative_area_level_1: window.lastSelectedGooglePlacesAddress.administrative_area_level_1,
-      administrative_area_level_2: window.lastSelectedGooglePlacesAddress.administrative_area_level_2,
-      postal_code: window.lastSelectedGooglePlacesAddress.postal_code,
-      country: window.lastSelectedGooglePlacesAddress.country,
-      geocoords: makePoint(window.lastSelectedGooglePlacesAddress.lon, window.lastSelectedGooglePlacesAddress.lat),
-    },
-    nameData: { first_name, last_name, voter_status },
-    user,
-  })(dispatch)
-}
-
-const signupAndEndorse = ({ vote, ...form }, state) => (dispatch) => {
-  const { location } = state
-  const { address, email, voter_status } = form
-  const error = validateNameAndAddressForm(address, form.name)
-
-  if (error) return dispatch({ type: 'error', error })
-
-  const name_pieces = form.name.split(' ')
-  const first_name = name_pieces[0]
-  const last_name = name_pieces.slice(1).join(' ')
-
-  return signIn({
-    channel: 'endorsement',
-    email,
-    device_desc: location.userAgent || 'Unknown',
-    phone_number: null,
-    redirect_to: location.path,
-  })(dispatch).then((user) => {
-    if (user) {
-      return updateNameAndAddress({
-        addressData: {
-          address,
-          locality: window.lastSelectedGooglePlacesAddress.locality,
-          administrative_area_level_1: window.lastSelectedGooglePlacesAddress.administrative_area_level_1,
-          administrative_area_level_2: window.lastSelectedGooglePlacesAddress.administrative_area_level_2,
-          postal_code: window.lastSelectedGooglePlacesAddress.postal_code,
-          country: window.lastSelectedGooglePlacesAddress.country,
-          geocoords: makePoint(window.lastSelectedGooglePlacesAddress.lon, window.lastSelectedGooglePlacesAddress.lat),
-        },
-        nameData: { first_name, last_name, voter_status },
-        user,
-      })(dispatch)
-        .then(() => endorse(vote, user, null, form.is_public)(dispatch))
-        .then(() => fetchMeasure(vote.measure.short_id, state)(dispatch))
-    }
-  })
-}
-
-const endorsementPageTitleAndMeta = (measures, vote, location) => {
-  const measure = measures[vote.measure.short_id]
-  const isCity = measure.legislature_name.includes(',')
-  const anonymousName = `${measure.legislature_name === 'U.S. Congress' ? 'American' : (stateNames[measure.legislature_name] || measure.legislature_name)} Resident`
-
-  const firstRealSentence = escapeHtml(vote.comment, { replaceApos: false, stripImages: true })
-    .split('\n').filter(line => line)[0] // strip leading whitespace
-  const title = `${vote.fullname || anonymousName}: ${firstRealSentence}`
-
-  // Determine social media image: start with database image, then any image in the comment, any image in measure summary, author avatar, legislature image
-  const dbImage = measure.image_name ? `${ASSETS_URL}/measure-images/${measure.image_name}` : ''
-  const inlineImageMatch = vote && vote.comment && vote.comment.match(/\bhttps?:\/\/\S+\.(png|jpg|jpeg|gif)\b/i)
-  const inlineImage = inlineImageMatch && inlineImageMatch[0]
-  const measureInlineImageMatch = measure && measure.summary && measure.summary.match(/\bhttps?:\/\/\S+\.(png|jpg|jpeg|gif)\b/i)
-  const measureInlineImage = measureInlineImageMatch && measureInlineImageMatch[0]
-  const authorImage = vote.username || vote.twitter_username ? avatarURL(vote) : null
-  const legislatureImage = (!isCity) ? `${ASSETS_URL}/legislature-images/${measure.legislature_name}.png` : ''
-
-  return {
-    ...location,
-    ogImage: dbImage || inlineImage || measureInlineImage || authorImage || legislatureImage,
-    ogTitle: `${measure.title} | Liquid US`,
-    description: title,
-    title,
-  }
-}
-
-const debounce = (function debouncer() {
-  let timeout = null
-  return (fn, ms) => {
-    if (timeout) clearTimeout(timeout)
-    timeout = setTimeout(fn, ms)
-  }
-}())
-
-const scrollQuestionFormIntoView = () => {
-  const elem = document.getElementById('measure-vote-form')
-    if (elem) {
-      const scrollY = elem.getBoundingClientRect().top + window.scrollY
-      if (scrollY) {
-        window.scrollTo(0, scrollY, { behavior: 'smooth' })
-      }
-    }
-  }
