+++ views/join-form-candidate.js
@@ -1,10 +1,13 @@
 const { handleForm, html } = require('../helpers')
+const { icon } = require('@fortawesome/fontawesome-svg-core')
+const { faUser } = require('@fortawesome/free-solid-svg-icons/faUser')
+const { faExclamationTriangle } = require('@fortawesome/free-solid-svg-icons/faExclamationTriangle')
 
 module.exports = ({ error, loading }, dispatch) => {
 
   return html`
     <div class="has-text-centered">
-      <form class="box has-text-centered" method="POST" onsubmit="${handleForm(dispatch, { type: 'contactForm:submitCandidatePage', url: 'https://liquid.us/candidate' })}">
+      <form class="box has-text-centered" method="POST" onsubmit="${handleForm(dispatch, { type: 'contactForm:messageSent', url: 'https://liquid.us/candidate' })}">
         <div class="field has-text-left">
           <strong><label for="idea">What policy idea deserves to go viral?</label><strong>
         </div>
@@ -26,10 +29,10 @@ module.exports = ({ error, loading }, dispatch) => {
           <div class="${`control is-expanded has-icons-left ${error ? 'has-icons-right' : ''}`}">
             <input name="email" class="${`input ${error ? 'is-danger' : ''}`}" type="text" required placeholder="you@example.com" />
             <span class="icon is-small is-left">
-              <i class="fa fa-user"></i>
+              ${icon(faUser)}
             </span>
             ${error ? html`<span class="icon is-small is-right">
-              <i class="fa fa-warning"></i>
+              ${icon(faExclamationTriangle)}
             </span>` : ''}
             ${error ? html`<p class="help is-danger">${error.message}</p>` : ''}
           </div>
