+++ views/join-page.js
@@ -1,5 +1,9 @@
 const { html } = require('../helpers')
 const joinForm = require('./join-form')
+const { icon } = require('@fortawesome/fontawesome-svg-core')
+const { faUsers } = require('@fortawesome/free-solid-svg-icons/faUsers')
+const { faAddressCard } = require('@fortawesome/free-solid-svg-icons/faAddressCard')
+const { faCheckSquare } = require('@fortawesome/free-solid-svg-icons/faCheckSquare')
 
 module.exports = (state, dispatch) => {
   return html`
@@ -16,15 +20,15 @@ module.exports = (state, dispatch) => {
             <div class="column has-text-centered">
               <h4 class="title is-4">
                 <span class="has-text-grey-light">&#9312;</span><br /><br />
-                <i class="fa fa-users" aria-hidden="true"></i>
+                <span class="icon">${icon(faUsers)}</span><br />
                 Proxying
-              </a></h4>
+              </h4>
               <p>Choose optional personal representatives so your values will always be counted.</p>
             </div>
             <div class="column has-text-centered">
               <h4 class="title is-4">
                 <span class="has-text-grey-light">&#9313;</span><br /><br />
-                <i class="far fa-address-card" aria-hidden="true"></i>
+                <span class="icon">${icon(faAddressCard)}</span><br />
                 Verification
               </h4>
               <p>Confirm your identity to ensure 1-person-1-vote.</p>
@@ -32,7 +36,7 @@ module.exports = (state, dispatch) => {
             <div class="column has-text-centered">
               <h4 class="title is-4">
                 <span class="has-text-grey-light">&#9314;</span><br /><br />
-                <i class="far fa-check-square" aria-hidden="true"></i>
+                <span class="icon">${icon(faCheckSquare)}</span><br />
                 Legislation
               </h4>
               <p>Vote directly on bills to hold your elected reps accountable.</p>
