+++ views/navbar-search.js
@@ -1,4 +1,8 @@
 const { avatarURL, html } = require('../helpers')
+const { icon } = require('@fortawesome/fontawesome-svg-core')
+const { faSearch } = require('@fortawesome/free-solid-svg-icons/faSearch')
+const { faUsers } = require('@fortawesome/free-solid-svg-icons/faUsers')
+const { faLandmark } = require('@fortawesome/free-solid-svg-icons/faLandmark')
 
 module.exports = ({ loading, query, results, showResults }, dispatch) => {
   return html`
@@ -11,9 +15,7 @@ module.exports = ({ loading, query, results, showResults }, dispatch) => {
               onblur=${(event) => dispatch({ type: 'blurred', event })}
               onkeyup=${(event) => dispatch({ type: 'inputChanged', event })}
               class="input" placeholder="Search Bills or People" />
-            <span class="icon is-left">
-              <i class="fa fa-search"></i>
-            </span>
+            <span class="icon is-left">${icon(faSearch)}</span>
           </div>
           <button type="submit" class="is-hidden">Search</button>
         </div>
@@ -76,7 +78,7 @@ const userSearchResult = (result, dispatch) => {
       <div class="media is-marginless">
         <div class="media-left">
           <div class="image is-32x32" style="${{ display: 'flex', alignItems: 'center', justifyContent: 'center' }}">
-            <img src=${avatarURL(result)} class="round-avatar-img" />
+            <img src=${avatarURL(result)} class="is-rounded" />
           </div>
         </div>
         <div class="media-content">
@@ -85,7 +87,7 @@ const userSearchResult = (result, dispatch) => {
             <span class="has-text-grey is-size-7">@${username || twitter_username}</span>
           </div>
           <div class="${direct_proxy_count ? 'is-size-7 has-text-grey' : 'is-hidden'}">
-            <span class="icon is-small"><i class="fa fa-users is-small"></i></span>
+            <span class="icon is-small">${icon(faUsers)}</span>
             <span>Represents ${direct_proxy_count} ${direct_proxy_count === 1 ? 'person' : 'people'} directly</span>
           </div>
         </div>
@@ -121,7 +123,7 @@ const measureSearchResult = ({ id, author_username, number, type, title, legisla
             class="image is-32x32 has-text-grey"
             style="${{ alignItems: 'center', display: 'flex', justifyContent: 'center' }}"
           >
-            <i class="fa fa-landmark"></i>
+            ${icon(faLandmark)}
           </div>
         </div>
         <div class="media-content">
