+++ views/petition-signup-form.js
@@ -1,4 +1,9 @@
 const { handleForm, html } = require('../helpers')
+const { icon } = require('@fortawesome/fontawesome-svg-core')
+const { faUser } = require('@fortawesome/free-solid-svg-icons/faUser')
+const { faExclamationTriangle } = require('@fortawesome/free-solid-svg-icons/faExclamationTriangle')
+const { faMapMarkerAlt } = require('@fortawesome/free-solid-svg-icons/faMapMarkerAlt')
+const { faEnvelope } = require('@fortawesome/free-solid-svg-icons/faEnvelope')
 
 module.exports = (state, dispatch) => {
   const { loading, error, measure } = state
@@ -14,8 +19,8 @@ module.exports = (state, dispatch) => {
         <div class="control has-icons-left">
           <input name="name" autocomplete="off" class=${`input ${error && error.field === 'name' && 'is-danger'}`} placeholder="John Doe" required />
           ${error && error.field === 'name'
-            ? html`<span class="icon is-small is-left"><i class="fas fa-exclamation-triangle"></i></span>`
-            : html`<span class="icon is-small is-left"><i class="fa fa-user"></i></span>`
+            ? html`<span class="icon is-small is-left">${icon(faExclamationTriangle)}</span>`
+            : html`<span class="icon is-small is-left">${icon(faUser)}</span>`
           }
           ${error && error.field === 'name' ? html`<p class="help is-danger">${error.message}</p>` : ''}
         </div>
@@ -26,8 +31,8 @@ module.exports = (state, dispatch) => {
           <div class="${`control is-expanded has-icons-left ${error && error.field === 'email' ? 'has-icons-right' : ''}`}">
             <input name="email" class="${`input ${error && error.field === 'email' ? 'is-danger' : ''}`}" type="text" placeholder="you@example.com" required />
             ${error && error.field === 'email'
-              ? html`<span class="icon is-small is-left"><i class="fas fa-exclamation-triangle"></i></span>`
-              : html`<span class="icon is-small is-left"><i class="fa fa-user"></i></span>`
+              ? html`<span class="icon is-small is-left">${icon(faExclamationTriangle)}</span>`
+              : html`<span class="icon is-small is-left">${icon(faEnvelope)}</span>`
             }
             ${error && error.field === 'email' ? html`<p class="help is-danger">This email is invalid.</p>` : ''}
           </div>
@@ -38,8 +43,8 @@ module.exports = (state, dispatch) => {
         <div class="control has-icons-left">
           <input onconnected=${initGoogleMaps} class=${`input ${error && error.field === 'address' && 'is-danger'}`} autocomplete="off" name="address" id="address_autocomplete_sidebar" placeholder="185 Berry Street, San Francisco, CA 94121" />
           ${error && error.field === 'address'
-            ? html`<span class="icon is-small is-left"><i class="fa fas fa-exclamation-triangle"></i></span>`
-            : html`<span class="icon is-small is-left"><i class="fa fa-map-marker-alt"></i></span>`
+              ? html`<span class="icon is-small is-left">${icon(faExclamationTriangle)}</span>`
+              : html`<span class="icon is-small is-left">${icon(faMapMarkerAlt)}</span>`
           }
           ${error && error.field === 'address' ? html`<p class="help is-danger">${error.message}</p>` : ''}
         </div>
