+++ views/proxies-page.js
@@ -15,7 +15,6 @@ module.exports = (state, dispatch) => {
           && !user.phone_verified
           && proxies.length > 0 ? html`
           <div class="notification">
-            <span class="icon"><i class="fa fa-exclamation-triangle"></i></span>
             You haven't verified yet. <strong><a href="/get_started">Finish verification</a></strong> so your proxies can represent you.
           </div>
         ` : ''}
